3rd Person Updated and more
The third person camera now follows the player properly.

When the player dies, the camera mode does not revert back to default
anymore.

When switching between the cameras, the canvases now know what camera is
active and updates itself. Script name is "canvasManager.js" and is
attached to the Canvas and CanvasR objects.

Basic features added for exiting the game and re-centering the rift. R
key re-centers, Esc key exits. The script that handles this is located
on the "OVRCameraRig" object and called "riftStandardScripts.cs"