Updated Player Prefab
The player camera now works like it did before and a new updated prefab
has been made.

The reason I changed it back is because the player camera cannot be
parented to the player in 3rd person or else the camera will rotate with
the player causing a bit of motion sickness.